requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
graphql-core==3.2.3
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
strawberry-graphql==0.209.6
typing_extensions==4.8.0
uvicorn==0.23.2
